Fix typo in 'VissibleName'.
BIG WARNING:

I'm both new to this Go and this package.  I own a reMarkable2 tablet and I'm
willing to put some effort into improving parts of this package.

These first commits are meant for me to explore the code base and, at the same
time, learn a bit of Go.

You may immediately notice that I include a Makefile with a simple target to
run main.go.  Which might a clue about me being a go-newbie.