Remove usage of deprecated io/ioutil.
This is just exercising and exploring the landscape of this source code.  At
the same time, I'm reading a bit of Effective Go.

I did noticed some unchecked return after a call to 'ioutils.WriteFile' in
'config/config.go'.  I don't think it matters because the function SaveTokens
is externally-infallible.